An importnat Update in XSD schemas and the PRO file
A place with no restriction on its capacity can own an unlimited
number of tokens (k = infinity). By the way, all places have an unlimited
capacity by default (capacity = 0 in the constructor).

When a Petri Net is saved as PNML document, all places that have
an ulimited capacity are saved with the attribute capicity="0":

                <toolspecific tool="petrinet" version="1.0">
                    <placeCapacity capacity="0"/>
                </toolspecific>

But in anyElement.xsd, the type of this attribute is forced to be strictly
greater than 0 (xs:positiveInteger) which will cause a validation error
when you open a PNML document!

So, we have to set the type to "xs:nonNegativeInteger".


Signed-off-by: issam <iabdallah@yandex.com>