Avoid node names uniqueness contraint
The PNML standard doesn't impose a contraint on node names uniqueness.
In fact, only nodes and arcs IDs must be unique: Two places or transitions
can have the same name but not the same ID!

Because nodes names similarity can harden and compilicate the P/T Net
and the Graph interpretation, it's not recommanded to use the same name
with two shown nodes.

When a name is used two times or more, the user will be warned.


Signed-off-by: issam <issam@iabdallah-pc.bfi>